+++ sdk/@launchdarkly/observability-android/lib/src/main/kotlin/com/launchdarkly/observability/context/LDObserveLogging.kt
@@ -1,4 +1,4 @@
-package com.launchdarkly.observability.devlog
+package com.launchdarkly.observability.context
 
 import android.util.Log
