+++ sdk/@launchdarkly/observability-android/lib/src/main/kotlin/com/launchdarkly/observability/context/ObserveLogAdapter.kt
@@ -1,4 +1,4 @@
-package com.launchdarkly.observability.devlog
+package com.launchdarkly.observability.context
 
 /**
  * Logging adapter interface for the Observability SDK.
