--- sdk/@launchdarkly/observability-android/lib/src/main/kotlin/com/launchdarkly/observability/devlog/LDContextCompat.kt
@@ -1,23 +1,0 @@
-package com.launchdarkly.observability.devlog
-
-import com.launchdarkly.sdk.LDContext
-
-/**
- * Converts an [LDContext] to an [LDObserveContext] for use in the Observability SDK.
- *
- * Use this at integration boundaries where code receives an [LDContext] from the
- * LaunchDarkly Client SDK and needs to pass it into observability/session-replay APIs.
- */
-fun LDContext.toLDObserveContext(): LDObserveContext {
-    if (isMultiple) {
-        val subs = (0 until individualContextCount).map { i ->
-            getIndividualContext(i).toLDObserveContext()
-        }
-        return LDObserveContext.createMulti(*subs.toTypedArray())
-    }
-    val builder = LDObserveContext.builder(kind.toString(), key)
-        .anonymous(isAnonymous)
-    val n = getName()
-    if (n != null) builder.name(n)
-    return builder.build()
-}
